+++ test/helpers/TestAttributeHelper.cs
@@ -19,7 +19,7 @@ namespace GdUnit4
 		public void TestGetAttribute_AttributePresent()
 		{
 			var obj = new TestClass();
-			var attribute = AttributeHelper.GetAttribute<CommandAttribute>(obj);
+			var attribute = Reflection.GetAttribute<CommandAttribute>(obj);
 
 			AssertObject(attribute).IsNotNull();
 			AssertString(attribute.Name).IsEqual("test");
@@ -29,7 +29,7 @@ namespace GdUnit4
 		public void TestGetAttribute_AttributeNotPresent()
 		{
 			var obj = new TestClass();
-			var attribute = AttributeHelper.GetAttribute<OptionAttribute>(obj);
+			var attribute = Reflection.GetAttribute<OptionAttribute>(obj);
 
 			AssertObject(attribute).IsNull();
 		}
@@ -38,7 +38,7 @@ namespace GdUnit4
 		public void TestGetAttribute_AttributesPresent()
 		{
 			var obj = new TestClass();
-			var attributes = AttributeHelper.GetAttributes<ArgumentAttribute>(obj);
+			var attributes = Reflection.GetAttributes<ArgumentAttribute>(obj);
 
 			AssertObject(attributes).IsNotNull();
 			AssertInt(attributes.Length).IsEqual(3);
@@ -51,7 +51,7 @@ namespace GdUnit4
 		public void TestGetAttribute_AttributesNotPresent()
 		{
 			var obj = new TestClass();
-			var attributes = AttributeHelper.GetAttributes<OptionAttribute>(obj);
+			var attributes = Reflection.GetAttributes<OptionAttribute>(obj);
 
 			AssertArray(attributes).IsEmpty();
 		}
@@ -60,7 +60,7 @@ namespace GdUnit4
 		public void TestHasAttribute_AttributePresent()
 		{
 			var obj = new TestClass();
-			var hasAttribute = AttributeHelper.HasAttribute<CommandAttribute>(obj);
+			var hasAttribute = Reflection.HasAttribute<CommandAttribute>(obj);
 
 			AssertBool(hasAttribute).IsEqual(true);
 		}
@@ -69,7 +69,7 @@ namespace GdUnit4
 		public void TestHasAttribute_AttributeNotPresent()
 		{
 			var obj = new TestClass();
-			var hasAttribute = AttributeHelper.HasAttribute<OptionAttribute>(obj);
+			var hasAttribute = Reflection.HasAttribute<OptionAttribute>(obj);
 
 			AssertBool(hasAttribute).IsEqual(false);
 		}